Parser completed

+++ src/main/scala/classes/Actor.scala
@@ -1,3 +1,9 @@
 package classes
 
-case class Actor(id: Int,login: String,display_login: String,gravatar_id: String,url: String,avatar_url: String)
+case class Actor(
+  id: BigInt,
+  login: String,
+  display_login: String,
+  gravatar_id: String,
+  url: String,
+  avatar_url: String)
