+++ src/main/scala/classes/Repo.scala
@@ -1,3 +1,3 @@
 package classes
 
-case class Repo(id: Int,name: String,url: String)
+case class Repo(id: BigInt,name: String,url: String)
